requirements.txt
Django==2.0.1
django-crispy-forms==1.7.0
django-filebrowser-no-grappelli==3.7.3
django-tinymce4-lite==1.7.0
jsmin==2.2.2
Pillow==5.0.0
python-decouple==3.1
pytz==2017.3
